cart controller + ref

+++ src/Tests/Demo/WebApi/OnlineStore/Drill4Net.Demo.OnlineStore.Dal/Models/CartItem.cs
@@ -7,7 +7,7 @@ namespace Drill4Net.Demo.OnlineStore.Dal.Models
         public Guid ProductId { get; set; }
         public string ProductName { get; set; }
         public decimal ProductPrice { get; set; }
-        public decimal ProductQuantity { get; set; }
+        public int ProductQuantity { get; set; }
         public decimal TotalPrice { get; set; }
     }
 }
