+++ src/Tests/Demo/WebApi/OnlineStore/Drill4Net.Demo.OnlineStore.WebApi/Models/CartDto.cs
@@ -9,6 +9,6 @@ namespace Drill4Net.Demo.OnlineStore.WebApi.Models
     {
         public Guid Id { get; set; }
         public List<CartItemDto> Products { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
     }
 }
